Separate cluster ops file, make logic work
